move metadataBuilder to common utils file
